feat: wip implement DiscoverContext
