+++ src/contexts/discovery/RepositoryConfig.ts
@@ -1,5 +1,9 @@
-interface RepositoryConfig {
-  remoteUrl: string;
-  host: string;
-  protocol: string;
+import { ServiceType, RemoteService, AccessType } from '../../detectors';
+
+export interface RepositoryConfig {
+  remoteService: RemoteService;
+  accessType: AccessType | undefined;
+  inputType: ServiceType | undefined;
+  host: string | undefined;
+  protocol: string | undefined;
 }
