+++ src/contexts/discovery/discoveryContextBinding.ts
@@ -1,21 +1,22 @@
 import { Container } from 'inversify';
 import { DiscoveryContextFactory, Types } from '../../types';
-import { RepositoryConfig } from './RepositoryConfig';
+import { RepositoryConfig } from '../discovery/RepositoryConfig';
+import { DiscoveryContext } from './DiscoveryContext';
 
 export const bindDiscoveryContext = (container: Container) => {
   container.bind(Types.DiscoveryContextFactory).toFactory(
     (): DiscoveryContextFactory => {
-      return (url: string) => {
-        const discoveryContextContainer = createDiscoveryContainer(url, container);
-        return discoveryContextContainer.get(RepositoryConfig);
+      return (uri: string) => {
+        const discoveryContextContainer = createDiscoveryContainer(uri, container);
+        return discoveryContextContainer.get(DiscoveryContext);
       };
     },
   );
 };
 
-const createDiscoveryContainer = (url: string, rootContainer: Container): Container => {
+const createDiscoveryContainer = (uri: string, rootContainer: Container): Container => {
   const container = rootContainer.createChild();
-  container.bind(Types.RepositoryConfig).toConstantValue(url);
+  container.bind(Types.RepositoryConfig).toConstantValue(uri);
 
   return container;
 };
