+++ src/types.ts
@@ -6,6 +6,8 @@ import { LanguageContext } from './contexts/language/LanguageContext';
 import { IProjectComponentDetector } from './detectors/IProjectComponentDetector';
 import { PracticeContext } from './contexts/practice/PracticeContext';
 import { ProjectComponentContext } from './contexts/projectComponent/ProjectComponentContext';
+import { RepositoryConfig } from './contexts/discovery/RepositoryConfig';
+import { DiscoveryContext } from './contexts/discovery/DiscoveryContext';
 
 export const Types = {
   //NEW TYPES
@@ -20,7 +22,9 @@ export const Types = {
   DiscoveryContextFactory: Symbol('Factory<DiscoveryContext>'),
   //Useful constants
   //Discovery context level
+
   RepositoryConfig: Symbol('RepositoryConfig'),
+
   //Scanner context level
   ScanningStrategy: Symbol('ScanningStrategy'),
   FileInspectorBasePath: Symbol('FileInspectorBasePath'),
@@ -57,4 +61,4 @@ export type LanguageContextFactory = (languageAtPath: LanguageAtPath) => Languag
 export type ProjectComponentContextFactory = (component: ProjectComponent) => ProjectComponentContext;
 export type ProjectComponentDetectorFactory = (language: ProgrammingLanguage) => IProjectComponentDetector[];
 export type PracticeContextFactory = (projectComponent: ProjectComponent) => PracticeContext;
-export type DiscoveryContextFactory = (owner: string, repo: string) => DiscoveryContext;
+export type DiscoveryContextFactory = (uri: string) => DiscoveryContext;
